// src/types/MovieInterface.ts
export interface MovieInterface {
    id: number;
    adult: boolean;
    backdrop_path: string;
    genre_ids: Array<number | undefined>;
    original_language: string;
    original_title: string;
    overview: string;
    popularity: number;
    poster_path: string;
    release_date: string;
    title: string;
    video: boolean;
    vote_average: number;
    vote_count: number
  }